requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
huggingface-hub>=0.20.0
requests>=2.31.0
pandas>=2.0.0
gradio>=4.0.0